docker-caas/requirements.txt
appnope==0.1.0
backports.ssl-match-hostname==3.4.0.2
decorator==4.0.4
Django==1.8.5
docker-py==1.4.0
gnureadline==6.3.3
ipython==4.0.0
ipython-genutils==0.1.0
MySQL-python==1.2.5
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
ptyprocess==0.5
requests==2.8.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.0.0
websocket-client==0.32.0
wheel==0.24.0